Dmirano/3786 activities restructure (#3958)
* layout restructure

* updated tests

* dashboard refactor

* update tests

* added login refactor

* updated tests

* added admin refactor

* updated tests

* route refactor

* route refactor to /pages/

* updated tests (not finishes yet)

* apd components refactor

* updated tests (route still causing issues)

* updated tests

* update tests, all pass now

* fixed merge conflicts

* relocated apd overview components

* updated tests and renamed to ApdOverview

* key state refactor

* updated tests

* previous activity restructure

* updated tests

* schedule summary refactor

* update tests

* executive summary restructure

* updated tests

* proposed budget refactor

* updated tests

* assurance and compliance restructure

* updated tests

* update tests after merge with proposed budget refactor

* export refactor

* updated tests

* activities refactor

* update tests

* updated failing tests
