+++ web/src/containers/activity/EntryPage.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import { useParams } from 'react-router-dom';
 
 import AlertMissingFFY from '../../components/AlertMissingFFY';
-import ActivityRoutes from './ActivityRoutes';
+import ActivityRoutes from '../../pages/apd/activities/ActivityRoutes';
 
 const EntryPage = ({ activityNames }) => {
   const activityIndex = +useParams().activityIndex;
@@ -12,7 +12,8 @@ const EntryPage = ({ activityNames }) => {
   return (
     <div className="remove-clearfix" id="activity-entry-page">
       <h2>
-        Activity {activityIndex + 1}: {activityNames[activityIndex] || "Untitled"}
+        Activity {activityIndex + 1}:{' '}
+        {activityNames[activityIndex] || 'Untitled'}
       </h2>
       <AlertMissingFFY />
       <ActivityRoutes activityIndex={activityIndex} />
